+++ event-trader/src/main/java/com/jgoetsch/eventtrader/order/size/OrderSize.java
@@ -15,11 +15,13 @@
  */
 package com.jgoetsch.eventtrader.order.size;
 
+import java.math.BigDecimal;
+
 import com.jgoetsch.eventtrader.TradeSignal;
 import com.jgoetsch.eventtrader.processor.ProcessorContext;
 
 public interface OrderSize {
 
-	public int getValue(TradeSignal trade, double price, ProcessorContext context);
+	public int getValue(TradeSignal trade, BigDecimal price, ProcessorContext context);
 
 }
